### 2020-08-14 #2575 @whedon
[REVIEW]: DYCO: A Python package to dynamically detect and compensate for time lags in ecosystem time series
**Submitting author:** @holukas (<a href="http://orcid.org/0000-0002-5569-0761">Lukas Hörtnagl</a>)
**Repository:** <a href="https://github.com/holukas/dyco" target ="_blank">https://github.com/holukas/dyco</a>
**Version:** v1.1.2
