### 2020-08-25 #2575 @holukas
@mccabete @KristinaRiemer I will check with ETH Zurich how they handle this and if they can open the registration.
