### 2020-08-25 #2575 @holukas
@mccabete @KristinaRiemer I just got feedback from ETH. While the code itself is open to everyone, issues can only be added after registering an account with an email address from the ETH domain. However, there is the possibility to create a guest account for the purpose of adding issues. Would this still be in line with JOSS requirements?
