### 2020-08-26 #2575 @KristinaRiemer
@holukas if it's possible to use guest accounts to create issues in the repository, and the issues and their comments are able to be seen publicly, I think that would be the best route. Let me know if that ends up not being possible though!
