### 2020-10-26 #2575 @mccabete
Sorry about the delay!

I am still running into issues - @holukas convinced ETH to give me a guest account (thanks!). However, it seems like ETH gave me an account that may/might not include github.ethz.ch privileges. I sent out an email to ETH Zurich's service desk today.
