### 2020-11-09 48887b6
Added logo.png
