### 2020-11-09 #2575 @holukas
Dear @KristinaRiemer @r-barnes @mccabete, after some back-and-forth I decided to move the code over to GitHub to make collaborations and the review work easier. I agree with @r-barnes that submitting issues here might get lost and is not ideal. @KristinaRiemer I would therefore like to cancel this current submission. I will then re-submit the same code, but this time using the GitHub repo. I hope the reviewers are still on-board to review the new submission. I am sorry for the inconvenience and hope for a smooth GitHub workflow. I already uploaded the code to https://github.com/holukas/dyco and I will make an official re-submission after this current one is closed. Thanks!
