### 2021-01-05 #2575 @mccabete
Thanks @KristinaRiemer  for the reminders. Sorry for not getting back to you both! @holukas, I added my comments as issues to the new repo.

The Wiki link is now behind an ETH login, so there was some content I had access to that I no longer have access to. That said, I had already gone through the wiki content before the migration, so I was able to post my comments. Likely 98% of my review is now posted, and once the link is fixed I can verify the final remaining check marks.

Thank you for being helpful and patient @KristinaRiemer  and @holukas
