### 2021-01-17 6a50c71
Added specific versions in setup.py
