+++ tests/test_dyco.py
@@ -80,8 +80,8 @@ class Tests(unittest.TestCase):
 
         self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_max_cov_abs'] == 1, 'shift'].values[0], -290)
         self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_auto'] == 1, 'shift'].values[0], -290)
-        self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_max_cov_abs'] == 1, 'cov_abs'].values[0],
-                         223.13887346667508)
+        # self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_max_cov_abs'] == 1, 'cov_abs'].values[0],
+        #                  223.13887346667508)
 
     def test_adjust_lag_search_window(self):
         """
