+++ setup.py
@@ -25,7 +25,7 @@ setuptools.setup(
     name='dyco',
     packages=setuptools.find_packages(),
     # packages=['dyco'],
-    version='1.0.2',
+    version='1.0.6',
     license='GNU General Public License v3 (GPLv3)',
     description='A Python package to detect and compensate for shifting lag times in ecosystem time series',
     long_description=long_description,
@@ -36,7 +36,7 @@ setuptools.setup(
     download_url='https://pypi.org/project/dyco/',
     keywords=['ecosystem', 'eddy covariance', 'fluxes',
               'time series', 'lag', 'timeshift'],
-    install_requires=['pandas==1.0.3', 'numpy==1.18.1', 'matplotlib==3.1.3', 'scipy==1.4.1'],
+    install_requires=['pandas==1.0.3', 'matplotlib==3.1.3', 'numpy==1.18.1', 'scipy==1.4.1'],
     classifiers=[
         'Development Status :: 4 - Beta',
         # Chose either "3 - Alpha", "4 - Beta" or "5 - Production/Stable" as the current state of your package
