+++ requirements.txt
@@ -17,6 +17,7 @@ dash-html-components==1.0.3
 dash-renderer==1.4.1
 dash-table==4.7.0
 docutils==0.16
+dyco==1.0.2
 entrypoints==0.3
 Flask==1.1.2
 Flask-Compress==1.5.0
