### 2021-01-22 holukas/dyco#4 @r-barnes
Consider putting README images in a subdir
As above, to increase organization.

openjournals/joss-reviews/issues/2575
