### 2021-01-22 holukas/dyco#5 @r-barnes
Test data description
A few README files sprinkled in the `test_data` folder describing the input data and its provenance would be useful.

openjournals/joss-reviews/issues/2575
