### 2021-01-22 holukas/dyco#6 @r-barnes
Friendlier wiki
The link to the wiki from the README drops onto a page from which it's not obvious where the user should go. I think it would be better if the README dropped the user onto a friendly landing page in the wiki from which it would be clear where to start.

In particular the page named "DYCO Wiki" appears to just duplicate the existing landing page.

openjournals/joss-reviews/issues/2575
