### 2021-01-22 holukas/dyco#7 @r-barnes
"Example settings"
On https://github.com/holukas/dyco/wiki/Example it's unclear that I'm supposed to click the "Example Settings" link to view the code being discussed. It would be better to either embed a copy of the code in the page or at least highlight that the user should open the link for cross-referencing.

openjournals/joss-reviews/issues/2575
