### 2021-01-22 holukas/dyco#8 @r-barnes
Pathlib
I see this
> At the moment it is also necessary to include the import from pathlib import Path.
in https://github.com/holukas/dyco/wiki/Example

Why is this so? Can it be fixed?

openjournals/joss-reviews/issues/2575
