### 2021-01-22 holukas/dyco#9 @r-barnes
Inputs/outputs
I feel like it's a little unclear to me from the README what the inputs and outputs of the software are.

Is it possible to summarize very briefly what it does? I feel like there's currently a wall-of-text issue.

openjournals/joss-reviews/issues/2575
