### 2021-01-22 holukas/dyco#10 @r-barnes
Command-line usage
The description [here](https://github.com/holukas/dyco/wiki/Processing-Chain) indicates that DYCO creates a number of folders and output files.

I think this isn't really appropriate for a tool that's expected to be used within Python. If this is a library, it's outputs should be variables that can be used programmatically. If this isn't a library, then there should be a command-line way to interact with the script.

openjournals/joss-reviews/issues/2575
