### 2021-01-22 holukas/dyco#11 @r-barnes
Test commented out
I see in `test_dyco.py` that at least one test is commented out:
```

    # def test_detect_covariance_peaks(self):
    #     """Test peak detection only"""
    #     filepath = 'test_data/test_raw_data/20161020113000.csv'
    #     segment_df = files.read_raw_data(filepath=filepath, data_timestamp_format='%Y-%m-%d %H:%M:%S.%f')
    #     lagsearch_df = lag.LagSearch.setup_lagsearch_df(win_lagsearch=[-1000, 1000],
    #                                                     shift_stepsize=10,
    #                                                     segment_name='20161031230000_iter1')
    #     lagsearch_df = \
    #         lag.LagSearch.find_max_cov_peak(segment_df=segment_df,
    #                                         lagsearch_df=lagsearch_df,
    #                                         ref_sig='w_ms-1_rot_turb',
    #                                         lagged_sig='co2_ppb_qcl_turb')
    #     lagsearch_df, props_peak_auto = lag.LagSearch.find_peak_auto(df=lagsearch_df)
    #
    #     self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_max_cov_abs'] == 1, 'shift'].values[0], -290)
    #     self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_auto'] == 1, 'shift'].values[0], -290)
    #     self.assertEqual(lagsearch_df.loc[lagsearch_df['flag_peak_max_cov_abs'] == 1, 'cov_abs'].values[0],
    #                      223.13887346667508)
```
Why is this? Please either fix the test or remove the dead code.

openjournals/joss-reviews/issues/2575
