### 2021-01-22 holukas/dyco#12 @r-barnes
Contributors
Please add a `CONTRIBUTING.md` file or similar.

openjournals/joss-reviews/issues/2575
