### 2021-01-22 holukas/dyco#13 @r-barnes
SUMMARY
I feel that reading the summary at https://github.com/holukas/dyco/blob/main/paper/paper.md does not give me a clear idea of what DYCO does. The summary mixes together background information and summary about the program.

I feel it would be clearer to state the background info in one paragraph and then explain in summary how DYCO solves the problem.

openjournals/joss-reviews/issues/2575
