### 2021-01-22 holukas/dyco#14 @r-barnes
State of the Field
The paper (https://github.com/holukas/dyco/blob/main/paper/paper.md) doesn't describe alternative methods or contrast DYCO's method with them. I think giving users a good sense of the trade-offs between various alternatives is important both for communicating the value of your work and for helping users make analysis choices appropriate to their situation.

openjournals/joss-reviews/issues/2575
