### 2021-01-22 holukas/dyco#15 @r-barnes
Citation reference format
I'm not sure that the references in https://github.com/openjournals/joss-reviews/issues/2575 follow the format JOSS wants. Please double-check this.

openjournals/joss-reviews/issues/2575
