### 2021-01-22 holukas/dyco#16 @r-barnes
Dead code
Please make sure there is no dead code. Either put code in a switch or remove it. Don't leave large sections commented-out. One example of the issue is [here](https://github.com/holukas/dyco/blob/main/dyco/analyze.py#L354-L377); there are others.

openjournals/joss-reviews/issues/2575
