### 2021-01-22 holukas/dyco#17 @r-barnes
Method/Function Documentation
There are a number of methods and functions without documentation in, e.g., these files:
https://github.com/holukas/dyco/blob/main/dyco/analyze.py
https://github.com/holukas/dyco/blob/main/dyco/plot.py
https://github.com/holukas/dyco/blob/main/dyco/files.py

Please make sure that all public methods/functions are appropriately documented.

openjournals/joss-reviews/issues/2575
