### 2021-01-22 #2575 @r-barnes
I've finished my review, per the above.

I think the biggest issue is that I'm left confused about whether Dyco is a tool I use kind of from the command-line or a library I use from within Python.

It looks as though Dyco produces a number of output directories and files, even if it's run from within Python. I think that's weird. Running commands from within a programming language should (mostly) give me products I can work with from within that language.

Conversely, the example I saw shows Dyco being run as a single, very large command from within Python to produce all of those outputs. So it's almost as though I have to open Python to run what's effectively a command-line command. If this is what's happening, then the authors should embrace the command-line nature of the work and modify their `setup.py` to add CLI commands.

Other than that, things look pretty good. The code appears clean and decently structured, though there is some missing documentation and dead code. Cleaning that up shouldn't be too difficult, though.
