### 2021-01-24 d16db2d
Removed dead code
