### 2021-01-24 holukas/dyco#16 @holukas
Removed dead code in d16db2df4ae43a0458e370f9bc61f7b32f29a097
