### 2021-01-27 afd1902
Implemented command line interface
