+++ dyco/cli.py
@@ -1,0 +1,113 @@
+"""
+    DYCO Dynamic Lag Compensation
+    Copyright (C) 2020  holukas
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+
+import argparse
+from pathlib import Path
+
+
+def validate_args(args):
+    """Check validity of optional args"""
+    if args.limitnumfiles < 0:
+        raise argparse.ArgumentTypeError("LIMITNUMFILES must be 0 or a positive integer.")
+    if args.lsnumiter < 1:
+        raise argparse.ArgumentTypeError("LSNUMITER must be > 1.")
+    if (args.lspercthres < 0.1) | (args.lspercthres > 1):
+        raise argparse.ArgumentTypeError("LSPERCTHRES must be between 0.1 and 1.")
+    if args.lssegmentduration > args.fileduration:  # todo
+        raise argparse.ArgumentTypeError("LSSEGMENTDURATION must be shorter or equal to FILEDURATION.")
+    if not args.lssegmentduration:
+        # If not specified, then lag times are determined using all of the file data
+        args.lssegmentduration = args.fileduration
+    if args.lsnumiter <= 0:
+        raise argparse.ArgumentTypeError("LSNUMITER must be a positive integer.")
+    args.lsremovefringebins = True if args.lsremovefringebins == 1 else False  # Translate settings to bool
+    args.delprevresults = True if args.delprevresults == 1 else False  # Translate settings to bool
+    return args
+
+
+def get_args():
+    """Get args from CLI input"""
+    parser = argparse.ArgumentParser(description="BICO - Conversion of ETH binary files to ASCII",
+                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+
+    # Positional args
+    parser.add_argument('var_reference', type=str,
+                        help="Column name of the unlagged reference variable in the data files (one-row header). "
+                             "Lags are determined in relation to this signal.")
+    parser.add_argument('var_lagged', type=str,
+                        help="Column name of the lagged variable in the data files (one-row header). "
+                             "The time lag of this signal is determined in relation to the reference "
+                             "signal var_reference.")
+    parser.add_argument('var_target', nargs='+',
+                        help="Column name(s) of the target variable(s). "
+                             "Column names of the variables the lag that was found between "
+                             "var_reference and var_lagged should be applied to. "
+                             "Example: var1 var2 var3")
+
+    # Optional args
+    parser.add_argument('-i', '--indir', type=Path,
+                        help="Path to the source folder that contains the data files, e.g. 'C:/bico/input'")
+    parser.add_argument('-o', '--outdir', type=Path,
+                        help="Path to output folder, e.g. C:/bico/output")
+    parser.add_argument('-fnd', '--filenamedateformat', type=str, default='%Y%m%d%H%M%S',
+                        help="Filename date format as datetime format strings. Is used to parse the date and "
+                             "time info from the filename of found files. The filename(s) of the files found in "
+                             "INDIR must contain datetime information. Example for data files named like "
+                             "20161015123000.csv: %%Y%%m%%d%%H%%M%%S")
+    parser.add_argument('-fnp', '--filenamepattern', type=str, default='*.csv',
+                        help="Filename pattern for raw data file search, e.g. *.csv")
[…]
