+++ setup.py
@@ -23,9 +23,9 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name='dyco',
+    entry_points={'console_scripts': ['dyco=dyco.cli:main']},
     packages=setuptools.find_packages(),
-    # packages=['dyco'],
-    version='1.0.6',
+    version='1.1.0-indev',
     license='GNU General Public License v3 (GPLv3)',
     description='A Python package to detect and compensate for shifting lag times in ecosystem time series',
     long_description=long_description,

[…]
