### 2021-01-27 holukas/dyco#10 @holukas
I agree and adjusted the code so it can be controlled from the command-line (afd19023c21ec49566d2fde5518a54d4e5c22f91).
