### 2021-01-28 a27f527
Added missing documentation for methods and functions
