+++ dyco/lag.py
@@ -105,7 +105,20 @@ class LagSearch:
         return None
 
     def flag_instantaneous_default_lag(self, default, cov_df):
-        """Flags location with instantaneous default time lag"""
+        """
+        Flags location with instantaneous default time lag
+
+        Parameters
+        ----------
+        default: int
+            Default lag time, given as record number
+        cov_df: pandas DataFrame
+            Lag search results for segment
+
+        Returns
+        -------
+        pandas DataFrame with flag True for default time lag location
+        """
         if not self.phase == 3:
             return cov_df
         default_ix = cov_df['shift'].loc[cov_df['shift'] == default].index[0]
@@ -114,6 +127,22 @@ class LagSearch:
 
     @staticmethod
     def setup_lagsearch_df(lgs_winsize: list, shift_stepsize: int, segment_name):
+        """
+        Setup DataFrame that collects lagsearch results
+
+        Parameters
+        ----------
+        lgs_winsize: list
+            Contains min and max of time lag search window, given as number of records
+        shift_stepsize: int
+            Stepsize used for shifting the lag search window
+        segment_name: str
+            Name the current lag search segment
+
+        Returns
+        -------
+        pandas DataFrame prepared for storing segment lag search results
+        """
         df = pd.DataFrame(columns=['index', 'segment_name', 'shift', 'cov', 'cov_abs',
                                    'flag_peak_max_cov_abs', 'flag_peak_auto'])
         df['shift'] = range(int(lgs_winsize[0]),
@@ -195,7 +224,23 @@ class LagSearch:
     def find_max_cov_peak(segment_df, cov_df, var_lagged, var_reference):
         """
         Find maximum absolute covariance
+
+        Parameters
+        ----------
+        segment_df: pandas DataFrame
+            Segment raw data
+        cov_df: pandas DataFrame
+            Stores results of segment lag search
+        var_lagged: str
+            Column name of the lagged variable
+        var_reference: str
+            Column name of the unlagged variable
+
+        Returns
+        -------
+        pandas DataFrame with segment lag search results
         """
+
 
         _segment_df = segment_df.copy()
         _segment_df['index'] = _segment_df.index
@@ -233,6 +278,18 @@ class LagSearch:
         return cov_df
 
     def save_cov_data(self, cov_df):
+        """
+        Save segment lag search results to csv file
+
+        Parameters
+        ----------
+        cov_df: pandas DataFrame
+            Segment lag search results
+
+        Returns
+        -------
+        None
+        """
         outfile = f'{self.segment_name}_segment_covariance_iteration-{self.iteration}.csv'
         outpath = self.outdir_data / outfile
         cov_df.to_csv(f"{outpath}")
@@ -262,7 +319,18 @@ class LagSearch:
         return idx
 
     def save_cov_plot(self, fig):
-        """Save covariance plot for segment"""
+        """
+        Save covariance plot for segment to png
+
+        Parameters
+        ----------
+        fig: Figure
+            The plot that is saved
+
+        Returns
+        -------
+        None
+        """
         outfile = f'{self.segment_name}_segment_covariance_iteration-{self.iteration}.png'
         outpath = self.outdir_plots / outfile
         fig.savefig(f"{outpath}", format='png', bbox_inches='tight', facecolor='w',
@@ -318,12 +386,28 @@ class LagSearch:
         return fig
 
     def z_as_colored_lines(self, fig, ax, x, y, z):
-        # z values as colors
-        # From: https://matplotlib.org/3.1.1/gallery/lines_bars_and_markers/multicolored_line.html
-        # Create a set of line segments so that we can color them individually
-        # This creates the points as a N x 1 x 2 array so that we can stack points
[…]
