### 2021-01-28 holukas/dyco#17 @holukas
Thanks for pointing this out, I added additional documentation in a27f527b6e10f8ffd8a0744554bbd74e76da80e5.
