+++ _wiki/_github dyco.md
@@ -1,202 +1,0 @@
-In this example, we install DYCO from pip and then use it in a small example script to find the time lag between two signals. We then remove the found time lag also from three other signals.
-
-If you want to follow along with this example and run it on your computer, you can download the example data and example script [here](https://github.com/holukas/dyco/tree/main/example).
-
-# Installation
-
-You can install `DYCO` using pip.
-`pip install dyco`
-
-Alternatively, you can download the source code and create a Python conda environment to run the code in. The respective `environment.yml` is provided.  
-Or, after downloading the source code, use the `requirements.txt` file to install all needed depencies via pip.
-
-# Example Data
-
-This example uses 20Hz raw data of turbulent wind and the greenhouse gases CO2 (carbon dioxide), N2O (nitrous oxide) and CH4 (methane). Data were measured in 2016 at the research station Davos in Switzerland ([more info](https://www.swissfluxnet.ethz.ch/index.php/sites/ch-dav-davos/site-info-ch-dav/))
-
-We will use 10 raw data files with data from 24 Oct 2016, recorded between 11:00 and 15:30. Each of these files contains 30 minutes of raw data:   
-`20161024100000.csv  20161024110000.csv  20161024120000.csv  20161024130000.csv  20161024140000.csv`  
-`20161024103000.csv  20161024113000.csv  20161024123000.csv  20161024133000.csv  20161024143000.csv`
-
-Note that in practice, considerably more than 10 files would be used with `DYCO`. Normally the script should be run on a dataset that consists of at least several days of data, but for the purpose of this example 10 files are sufficient. For the research site Davos, `DYCO` is used on multi-year datasets.
-
-Each of these files contains more data than the variables we are interested in.
-
-Data in the files look like this (20161024110000.csv):
-* Header with variable names in 1st row
-* Timestamp in 1st column
-* Data starting in 2nd column
-
-In this example, the imporant variables are: 
-* wind (`w_ms-1_rot_turb`)
-* CO2 (`co2_ppb_qcl_turb`)
-* N2O (`n2o_ppb_q`
-* `cl`)
-* CH4 (`ch4_ppb_qcl`)
-
-# Goal
-
-The goal is to prepare the N2O and CH4 data in the raw data files for an easier and more accurate calculation of the respective ecosystem fluxes.
-
-First we want to remove the time lag between the wind variable and the CO2 variable. Then, we will use this information to remove the time shift also from the other signals (N2O, CH4). Finding the lag between wind and CO2 is relatively easy because the CO2 exchange at the site is strong. In contrast, finding the lag between wind and N2O or CH4 is challenging due to low N2O and CH4 exchange.
-
-# Running The Code
-
-DYCO is run from the command line interface (CLI).
-
-## Example with full arguments
-
-`python L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\dyco\dyco.py w_ms-1_rot_turb co2_ppb_qcl_turb n2o_ppb_qcl ch4_ppb_qcl -i L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\example\input_data -o L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\example\output -fnd %Y%m%d%H%M%S -fnp *.csv -flim 0 -fgr 30T -fdur 30T -dtf "%Y-%m-%d %H:%M:%S.%f" -dres 0.05 -lss 30T -lsw 1000 -lsi 3 -lsf 1 -lsp 0.9 -lt 0 -del 0`
-
-## Usage
-
[…]
