### 2021-01-28 7fce4ed
Adjusted citation style
