+++ paper/paper.bib
@@ -1,5 +1,5 @@
 
-@article{baldocchi_measuring_1988,
+@article{Baldocchi1988,
 	title = {Measuring {Biosphere}-{Atmosphere} {Exchanges} of {Biologically} {Related} {Gases} with {Micrometeorological} {Methods}},
 	volume = {69},
 	issn = {00129658},
@@ -17,7 +17,7 @@
 	file = {Baldocchi, Hincks, Meyers - 1988 - Measuring Biosphere-Atmosphere Exchanges of Biologically Related Gases with Micrometeorological Metho.pdf:L\:\\Dropbox\\luhk_work\\PDF\\PEER-REVIEWED\\B\\Baldocchi, Hincks, Meyers - 1988 - Measuring Biosphere-Atmosphere Exchanges of Biologically Related Gases with Micrometeorological Metho.pdf:application/pdf}
 }
 
-@article{langford_eddy-covariance_2015,
+@article{Langford2015,
 	title = {Eddy-covariance data with low signal-to-noise ratio: time-lag determination, uncertainties and limit of detection},
 	volume = {8},
 	issn = {1867-8548},
@@ -36,7 +36,7 @@
 	file = {Langford et al. - 2015 - Eddy-covariance data with low signal-to-noise ratio time-lag determination, uncertainties and limit of detectio.pdf:L\:\\Dropbox\\luhk_work\\PDF\\PEER-REVIEWED\\L\\Langford et al. - 2015 - Eddy-covariance data with low signal-to-noise ratio time-lag determination, uncertainties and limit of detectio.pdf:application/pdf}
 }
 
-@article{moncrieff_system_1997,
+@article{Moncrieff1997,
 	title = {A system to measure surface fluxes of momentum, sensible heat, water vapour and carbon dioxide},
 	volume = {188-189},
 	issn = {00221694},
@@ -53,7 +53,7 @@
 	file = {Moncrieff et al. - 1997 - A system to measure surface fluxes of momentum, sensible heat, water vapour and carbon dioxide.pdf:L\:\\Dropbox\\luhk_work\\PDF\\PEER-REVIEWED\\M\\Moncrieff et al. - 1997 - A system to measure surface fluxes of momentum, sensible heat, water vapour and carbon dioxide.pdf:application/pdf}
 }
 
-@article{merbold_greenhouse_2014,
+@article{Merbold2014,
 	title = {Greenhouse gas budget ({CO} $_{\textrm{2}}$ , {CH} $_{\textrm{4}}$ and {N} $_{\textrm{2}}$ {O}) of intensively managed grassland following restoration},
 	volume = {20},
 	issn = {13541013},
@@ -71,7 +71,7 @@
 	file = {Merbold et al. - 2014 - Greenhouse gas budget (CO2, CH4 and N2O) of intensively managed grassland following restoration.pdf:L\:\\Dropbox\\luhk_work\\PDF\\PEER-REVIEWED\\M\\Merbold et al. - 2014 - Greenhouse gas budget (CO2, CH4 and N2O) of intensively managed grassland following restoration.pdf:application/pdf}
 }
 
-@article{mcmillen_eddy_1988,
+@article{McMillen1988,
 	title = {An eddy correlation technique with extended applicability to non-simple terrain},
 	volume = {43},
 	issn = {0006-8314, 1573-1472},
@@ -89,7 +89,7 @@
[…]
