### 2021-01-28 holukas/dyco#15 @holukas
Seems like I forgot to follow the JOSS citation guidelines, I fixed this in 7fce4edf4446bb3f383eccf5dc90334b1338af90.
