### 2021-01-31 holukas/dyco#14 @holukas
I updated the paper and tried to re-word the _Statement of need_ section in a way so that it better reflects which unique features of DYCO can assist current flux processing schemes in calculating ecosystem fluxes (61ee45db749239be03bcadb8f2f7aa4d6e9fc123). To my knowledge, there is currently no openly available script or other software that offers similarly dynamic processing options for eddy covariance raw data.
