### 2021-01-31 holukas/dyco#13 @holukas
I restructured the paper and re-worded the _Summary_, which now contains some of the info previously found in _Scientific background_ in the first paragraph and then mentions how DYCO can help in addressing the challenges (61ee45db749239be03bcadb8f2f7aa4d6e9fc123).

The _Statement of need_ was also reworded and now goes into more detail regarding the unique features DYCO can offer to solve commonly encountered problems when dealing with low-signal-to-noise-ratio compounds.

In addition, I felt it would be helpful to give the reader a quick overview of the _Processing chain_ in a dedicated section.

I also updated the _Real-world examples_ with a short example how DYCO can help with the calculation of N2O fluxes over a managed grassland.
