### 2021-01-31 holukas/dyco#12 @holukas
Done, created `CONTRIBUTING.md` in c5faedc27bc820a376a06a705e2a54f762c433fd.
