+++ dyco/dyco.py
@@ -465,52 +465,52 @@ class DynamicLagCompensation:
 
 
 def main(args):
-    # DynamicLagCompensation(var_reference=args.var_reference,
-    #                        var_lagged=args.var_lagged,
-    #                        var_target=args.var_target,
-    #                        indir=args.indir,
-    #                        outdir=args.outdir,
-    #                        fnm_date_format=args.filenamedateformat,
-    #                        fnm_pattern=args.filenamepattern,
-    #                        files_how_many=args.limitnumfiles,
-    #                        file_generation_res=args.filegenres,
-    #                        file_duration=args.fileduration,
-    #                        dat_recs_timestamp_format=args.datatimestampformat,
-    #                        dat_recs_nominal_timeres=args.datanominaltimeres,
-    #                        lgs_segment_dur=args.lssegmentduration,
-    #                        lgs_winsize=args.lswinsize,
-    #                        lgs_num_iter=args.lsnumiter,
-    #                        lgs_hist_remove_fringe_bins=args.lsremovefringebins,
-    #                        lgs_hist_perc_thres=args.lspercthres,
-    #                        target_lag=args.targetlag,
-    #                        del_previous_results=args.delprevresults)
+    DynamicLagCompensation(var_reference=args.var_reference,
+                           var_lagged=args.var_lagged,
+                           var_target=args.var_target,
+                           indir=args.indir,
+                           outdir=args.outdir,
+                           fnm_date_format=args.filenamedateformat,
+                           fnm_pattern=args.filenamepattern,
+                           files_how_many=args.limitnumfiles,
+                           file_generation_res=args.filegenres,
+                           file_duration=args.fileduration,
+                           dat_recs_timestamp_format=args.datatimestampformat,
+                           dat_recs_nominal_timeres=args.datanominaltimeres,
+                           lgs_segment_dur=args.lssegmentduration,
+                           lgs_winsize=args.lswinsize,
+                           lgs_num_iter=args.lsnumiter,
+                           lgs_hist_remove_fringe_bins=args.lsremovefringebins,
+                           lgs_hist_perc_thres=args.lspercthres,
+                           target_lag=args.targetlag,
+                           del_previous_results=args.delprevresults)
 
-    # Example with full arguments
-    DynamicLagCompensation(
-        var_reference='w_ms-1_rot_turb',  # Reference signal
-        var_lagged='co2_ppb_qcl_turb',  # Lagged signal
-        var_target=['co2_ppb_qcl_turb', 'n2o_ppb_qcl', 'ch4_ppb_qcl'],  # Target columns
-        indir=Path('L:/Dropbox/luhk_work/programming/DYCO_Dynamic_Lag_Compensation/example/input_data'),
-        outdir=Path('L:/Dropbox/luhk_work/programming/DYCO_Dynamic_Lag_Compensation/example/output'),
-        fnm_date_format='%Y%m%d%H%M%S',  # Date format in filename
-        fnm_pattern='2016*.csv',  # Filename pattern for file search, accepts regex
-        # dat_recs_timestamp_format='%Y-%m-%d %H:%M:%S.%f',  # Timestamp format for each row record, will be created if False
-        dat_recs_nominal_timeres=0.05,  # Nominal (expected) time resolution, one record every x seconds
-        files_how_many=False,  # Limit number of files to use
-        file_generation_res='30T',  # One file generated every 30 minutes
-        file_duration='30T',  # Duration of one data file is 30 minutes
[…]
