+++ README.md
@@ -1,40 +1,106 @@
-![Normalization example](logo.png)
+![Normalization example](images/logo_dyco1_256px.png)
 
 # DYCO - Dynamic Lag Compensation
 
-`DYCO` facilitates the detection of time lags between two time series using covariance maximization. The time delay for each averaging interval is estimated by incrementally shifting one signal (the *reference signal*) and repeatedly calculating the time-lagged cross-covariance between both variables. The time lag that produces the highest covariance is selected as the *reference lag*, which can then be applied to selected *target* *variables*. The full processing chain comprises several phases and iterations during which *reference lags* are refined iteratively in progressively smaller search windows. The lag search is facilitated by prior normalization of default (nominal) time lags across files. This is achieved by compensating data for daily default lags, calculated from high-quality *reference lags* available around the respective date (Figure 1). Due to this normalization, *reference lags* fall into a specific, pre-defined and therefore known time range, which in turn allows the application of increasingly narrow time windows during lag search.
+The lag detection between the turbulent departures of measured wind and the scalar of interest is a central step in the calculation of eddy covariance (EC) ecosystem fluxes. In case the covariance maximization fails to detect a clear peak in the covariance function between the wind and the scalar, current flux calculation software can apply a constant default (nominal) time lag to the respective scalar. However, both the detection of clear covariance peaks in a pre-defined time window and the definition of a reliable default time lag is challenging for compounds which are often characterized by low SNR, such as N<sub>2</sub>O. In addition, the application of one static default time lag may produce inaccurate results in case systematic time shifts are present in the raw data.
 
-This approach has the advantage that *reference lags* can be calculated from signals that show clear peaks in the cross-covariance analysis and thus yield unambiguous time lags due to their high signal-to-noise ratio (SNR). In case the lag search failed to detect a clear time delay for the *reference variable*, the respective daily default lag is used instead.  *Reference lags* can then be used to compensate *target variables* with low SNR for detected *reference* time delays. 
+ `DYCO` is meant to assist current flux processing software in the calculation of fluxes for compounds with low SNR. In the context of current flux processing schemes, the unique features offered as part of the `DYCO` package include:
+
+- (i) the dynamic application of progressively smaller time windows during lag search for a *reference* compound (e.g. CO<sub>2</sub>),
+- (ii) the calculation of default time lags on a daily scale for the *reference* compound,
+- (iii) the application of daily default *reference* time lags to one or more *target* compounds (e.g. N<sub>2</sub>O)
+- (iv) the dynamic normalization of time lags across raw data files,
+- (v) the automatic correction of systematic time shifts in *target* raw data time series, e.g. due to failed synchronization of instrument clocks, and
+- (vi) the application of instantaneous *reference* time lags, calculated from lag-normalized files, to one or more *target* compounds.
+
+As `DYCO` aims to complement current flux processing schemes, final lag-removed files are produced that can be directly used in current flux calculation software.
 
 
 
-![Normalization example](paper/figure.png)
+## Scientific background
 
[…]
