### 2021-02-04 holukas/dyco#8 @holukas
After some thinking and then some more trial-and-error, I was able to remove the pathlib requirement in afd19023c21ec49566d2fde5518a54d4e5c22f91.
