### 2021-02-04 69a7ccb
Updated JOSS paper and README descriptions
