+++ example/example.py
@@ -11,24 +11,24 @@ https://github.com/holukas/dyco/wiki/Usage
 
 import os
 
-cli = 'python L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\dyco\dyco.py ' \
-      'w_ms-1_rot_turb co2_ppb_qcl_turb ' \
-      'co2_ppb_qcl_turb n2o_ppb_qcl ch4_ppb_qcl ' \
-      '-i L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\example\input_data ' \
-      '-o L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\example\output ' \
-      '-fnd %Y%m%d%H%M%S ' \
-      '-fnp *.csv ' \
-      '-flim 0 ' \
-      '-fgr 30T ' \
-      '-fdur 30T ' \
-      '-dtf "%Y-%m-%d %H:%M:%S.%f" ' \
-      '-dres 0.05 ' \
-      '-lss 30T ' \
-      '-lsw 1000 ' \
-      '-lsi 3 ' \
-      '-lsf 1 ' \
-      '-lsp 0.9 ' \
-      '-lt 0 ' \
-      '-del 0'
+cli = r'python L:\Dropbox\luhk_work\programming\DYCO_Dynamic_Lag_Compensation\dyco\dyco.py ' \
+      r'w_ms-1_rot_turb co2_ppb_qcl_turb ' \
+      r'co2_ppb_qcl_turb n2o_ppb_qcl ch4_ppb_qcl ' \
+      r'-i A:\FLUXES\CH-DAV\0___ms_CH-DAV_non-CO2\tests_dynamic_lag\JOSS_paper\1_FISP_output\splits ' \
+      r'-o A:\FLUXES\CH-DAV\0___ms_CH-DAV_non-CO2\tests_dynamic_lag\JOSS_paper\2_DYCO_output_2_noTimestampInPhase3 ' \
+      r'-fnd %Y%m%d%H%M%S ' \
+      r'-fnp *.csv ' \
+      r'-flim 0 ' \
+      r'-fgr 30T ' \
+      r'-fdur 30T ' \
+      r'-dtf "%Y-%m-%d %H:%M:%S.%f" ' \
+      r'-dres 0.05 ' \
+      r'-lss 30T ' \
+      r'-lsw 1000 ' \
+      r'-lsi 3 ' \
+      r'-lsf 1 ' \
+      r'-lsp 0.9 ' \
+      r'-lt 0 ' \
+      r'-del 0'
 print(cli)
 os.system(cli)
