### 2021-02-04 holukas/dyco#9 @holukas
I tried to clarify this point and adjusted the texts in the paper and the README (69a7ccb115e172e344aa8f92693273b4d4674c9d).

- The texts now mention eddy covariance raw data files specifically as input files:
> `DYCO` uses eddy covariance raw data files as input and produces lag-compensated raw data files as output.
- The `DYCO` [processing chain flowchart ](https://github.com/holukas/dyco/blob/main/paper/dyco_processing_chain.png)now mentions eddy covariance data files specifically as input.
