### 2021-02-04 954d825
Reactivated test for auto peak detection
