### 2021-02-04 holukas/dyco#11 @holukas
The test got deactivated due to a renaming issue, but I adjusted the code and reactivated the test in 954d8254f0dee915e7b9f75745262084c28bd000.
