### 2021-02-04 holukas/dyco#4 @holukas
I put all README images in the images subdir (6ed716be55311ba1e955ad9345dada68670ecd98).
