### 2021-02-04 holukas/dyco#6 @holukas
I removed the duplicate and tried to make the welcome page a bit nicer. In addition, all texts in the wiki were updated and more plots showing DYCO results were added.

https://github.com/holukas/dyco/wiki
