### 2021-02-04 #2575 @holukas
Hi @KristinaRiemer, I am nearly finished with the Issues. I was not sure if I should close the issues if I think they are done, or if it is up to the reviewers to confirm that an issue was solved and then I can close them. What would be the correct procedure?
