### 2021-02-04 #2575 @KristinaRiemer
@holukas I think we should do the latter. When you're done addressing all of the issues, let the reviewers know here. They can then review the work you've done and you can close the issues once they indicate they're satisfied with it. Hopefully that's the easiest way for @r-barnes and @mccabete to keep track of everything!
