### 2021-02-04 holukas/dyco#7 @holukas
I restructured and reworded the [Example](https://github.com/holukas/dyco/wiki/Example) and tried to include more information about how the script is used. I also added an example for the command line interface with full arguments and added separate wiki pages for [Usage](https://github.com/holukas/dyco/wiki/Usage) and [Results Output Folders](https://github.com/holukas/dyco/wiki/Results-Output-Folders).
