### 2021-02-04 holukas/dyco#5 @holukas
I added a description for each of the data subdirs in 81baee86fb3b8aff13fbdb89718969c92003cc01.
