### 2021-02-06 #2575 @holukas
Dear @KristinaRiemer @r-barnes @mccabete,

thank you all for your input and feedback, highly appreciated! I have now addressed all issues and followed up on your suggestions. The script and documentation underwent quite some changes and I hope DYCO now meets the criteria for publication. Please let me know in case there are further changes you would like to see.

Cheers
Lukas
