### 2021-02-19 #2575 @KristinaRiemer
Hi @r-barnes and @mccabete, have you had a chance yet to review the addressing of your issues in the [DYCO repo](https://github.com/holukas/dyco) yet? Please let me know if you need any help.
