### 2021-04-18 holukas/dyco#16 @r-barnes
I'm still seeing a bit of this in tests/test_dyco.py

I won't hold up the review any further on it, but could you check into the situation? It's ~5 LOC.
