### 2021-04-18 holukas/dyco#10 @r-barnes
Hey, if I don't mention it elsewhere, nice use of type annotations (at least in some places)!
