### 2021-04-18 holukas/dyco#17 @r-barnes
pylint still shows some undocumented functions/classes:

dyco/setup_dyco.py:97:0: C0115: Missing class docstring (missing-class-docstring)
dyco/setup_dyco.py:104:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/setup_dyco.py:195:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/dyco.py:52:4: C0115: Missing class docstring (missing-class-docstring)
dyco/dyco.py:467:0: C0116: Missing function or method docstring (missing-function-docstring)
dyco/plot.py:32:0: C0115: Missing class docstring (missing-class-docstring)
dyco/plot.py:44:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/plot.py:124:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/plot.py:169:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/plot.py:258:0: C0116: Missing function or method docstring (missing-function-docstring)
dyco/plot.py:284:0: C0116: Missing function or method docstring (missing-function-docstring)
dyco/loop.py:304:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/loop.py:583:0: C0115: Missing class docstring (missing-class-docstring)
dyco/lag.py:220:4: C0116: Missing function or method docstring (missing-function-docstring)
dyco/lag.py:515:0: C0115: Missing class docstring (missing-class-docstring)
dyco/lag.py:569:4: C0116: Missing function or method docstring (missing-function-docstring)
