### 2021-04-18 #2575 @r-barnes
I feel like my concerns have been reasonably addressed and that the remaining issues I've left open shouldn't preclude moving this toward publication. My ability to really deep-dive into the results/processing of the package is hampered by my lack of domain knowledge here, but I think there's sufficient explanatory information that a user with this knowledge would be able to figure out what's going on and derive value from this software. I'm happy to approve this.
