### 2021-04-26 0ff3f62
Added missing docstrings
