### 2021-04-26 holukas/dyco#17 @holukas
I added the missing docstrings in 0ff3f62627e40181df2c6d589ff2f8786d6596de. I hope I did not miss anything.
