### 2021-04-26 holukas/dyco#16 @holukas
Yes there is some more, I must have missed it. I removed the dead code in 9bd00c6e0c6949c04a62c2a5ba71ca47c4b4a588.
