### 2021-06-02 #2575 @KristinaRiemer
@arfon thank you for checking in on this! @mccabete it looks like the "functionality" and "summary" items in the check list are currently unchecked, do you have further comments about those areas?
