### 2021-06-04 #2575 @KristinaRiemer
Okay, thanks @mccabete. I checked those off just for the sake of completion.

@holukas there are a few more steps to go before publication. I'm going to check over the paper for typos and such, and it's suggested that you do also. Once that's done, I'll ask for archive links.
