### 2021-06-04 #2575 @KristinaRiemer
Okay, I found only a few small errors for you to fix in the paper @holukas:

- Choose either bullets or numerals for the list at the top of the second page
- The first sentence in line 51 appears to be incomplete? I'm not sure if it should be joined to the following sentence or if there's something missing from it
- For the two figures, they only need one figure caption each. You can remove the one that's included when inserting each image if you want
- In line 124, the word "approx." should be fully spelled out
- In the last reference, in International Agrophysics, there are a few authors whose names should be fully included instead of condensed as an ellipsis

It also looks like the subscripts on CO2 and N20 aren't rendering. You put them correctly in the markdown, so I'm not sure what's going on and someone is looking into why that's happening.
