### 2021-06-04 #2575 @KristinaRiemer
Hi @holukas, it looks like the syntax for the subscripts should be with `~` instead of `<sub>` for some reason. So `CO<sub>2</sub>` should instead be `CO~2~`. If you can modify those in the paper, we'll see if that works!
