### 2021-06-10 #2575 @holukas
Thanks @KristinaRiemer for checking the PDF. I made the corrections and adjustments, except regarding the reference International Agrophysics. It seems like that the ellipsis was added when the PDF was generated and I am not sure I can change this from my side. In the .bib file all authors are included with their full name. Regarding the double captions for the figures, I now left the captions empty when importing the images, I hope that works.
