### 2021-06-10 #2575 @KristinaRiemer
Okay, I can't figure out why it truncated the author list for that reference, but I guess it's a default for a reason, so we can leave it. Thanks for making all those changes to the paper!

The next steps for you to do @holukas are: 1) Create a tagged release of the GitHub repo (instructions [here](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository#creating-a-release)) and post the version number here, and 2) generate a DOI by archiving the code on Zenodo, figshare, an institutional repository, etc. and provide that DOI. Make sure the archive has the same title as the paper title, and the same authors.
