### 2021-06-13 bc2bf70
Updated version number for JOSS release
