### 2021-06-16 a6d8cb4
Added Zenodo DOI badge
