### 2021-06-16 #2575 @holukas
@KristinaRiemer
1) I have created a new release on GitHub, the version number is v1.1.2: https://github.com/holukas/dyco/releases/tag/v1.1.2
2) I have archived the this version 1.1.2 on Zenodo, DOI: 10.5281/zenodo.4964068
https://zenodo.org/record/4964068
