### 2021-06-16 #2575 @KristinaRiemer
@holukas thanks for getting these together!

For the Zenodo archive, could you change the name of the archive to the name of the paper ("DYCO: A Python package to dynamically detect and1compensate for time lags in ecosystem time series")? Additionally, could you make the author name your name as in the paper?
