### 2021-06-16 #2575 @danielskatz
@holukas - As part of a final proofread, I have suggested some changes in https://github.com/holukas/dyco/pull/19 - please merge these or let me know what you disagree with, then we can proceed with the publishing.
