+++ paper/paper.md
@@ -25,7 +25,7 @@ One method to adequately calculate fluxes for compounds with low SNR is to first
 
 # Statement of need
 
-Lag detection between the turbulent departures of measured wind and the scalar of interest is a central step in the calculation of EC ecosystem fluxes. In the covariance maximization fails to detect a clear peak in the covariance function between the wind and the scalar, current flux calculation software can apply a constant default (nominal) time lag to the respective scalar. However, both the detection of clear covariance peaks in a pre-defined time window and the definition of a reliable default time lag is challenging for compounds that are often characterized by low SNR, such as N~2~O. In addition, the application of one static default time lag may produce inaccurate results when systematic time shifts are present in the raw data.
+Lag detection between the turbulent departures of measured wind and the scalar of interest is a central step in the calculation of EC ecosystem fluxes. If the covariance maximization fails to detect a clear peak in the covariance function between the wind and the scalar, current flux calculation software can apply a constant default (nominal) time lag to the respective scalar. However, both the detection of clear covariance peaks in a pre-defined time window and the definition of a reliable default time lag is challenging for compounds that are often characterized by low SNR, such as N~2~O. In addition, the application of one static default time lag may produce inaccurate results when systematic time shifts are present in the raw data.
 
  `DYCO` is meant to assist current flux processing software in the calculation of fluxes for compounds with low SNR. In the context of current flux processing schemes, the unique features offered as part of the `DYCO` package include:
 
