### 2021-06-16 #2575 @holukas
@danielskatz There is only one small typo that I found. In the PDF L131 please change "compensated" to "compensates", rest looks good.
