### 2021-06-16 #2575 @danielskatz
You can and should do that, since it's in your repo :)  When you do it, run ```@whedon generate pdf``` again and then let me know that it's ready to publish
