+++ paper/paper.md
@@ -68,7 +68,7 @@ The full `DYCO` processing chain comprises four phases and several iterations du
 
 # Real-world examples
 
-The [ICOS](https://www.icos-cp.eu/) Class 1 site [Davos](https://www.swissfluxnet.ethz.ch/index.php/sites/ch-dav-davos/site-info-ch-dav/) (CH-Dav), a subalpine forest ecosystem station in the east of Switzerland, provides one of the longest continuous time series (24 years and running) of ecosystem fluxes globally. Since 2016, measurements of the strong GHG N~2~O are recorded by a closed-path gas analyzer that also records CO~2~. To calculate fluxes using the EC method, wind data from the sonic anemometer is combined with instantaneous gas measurements from the gas analyzer. However, the air sampled by the gas analyzer needs a certain amount of time to travel from the tube inlet to the measurement cell in the analyzer and is thus lagged behind the wind signal. The lag between the two signals needs to be compensated for by detecting and then removing the time lag at which the cross-covariance between the turbulent wind and the turbulent gas signal reaches the maximum absolute value. This works generally well when using CO~2~ (high SNR) but is challenging for N~2~O (low SNR). Using covariance maximization to search for the lag between wind and N~2~O mostly fails to accurately detect time lags between the two variables (noisy cross-correlation function), resulting in relatively noisy fluxes. However, since N~2~O has similar adsorption / desorption characteristics as CO~2~, it is valid to assume that both compounds need approximately the same time to travel through the tube to the analyzer, i.e., the time lag for both compounds in relation to the wind is similar. Therefore, `DYCO` can be applied (i) to calculate time lags across files for CO~2~ (*reference* compound), and then (ii) to remove found CO~2~ time delays from the N~2~O signal (*target* compound). The lag-compensated files produced by `DYCO` can then be used to calculate N~2~O fluxes. Since `DYCO` normalizes time lags across files and compensated the N~2~O signal for instantaneous CO~2~ lags, the *true* lag between wind and N~2~O can be found close to zero, which in turn facilitates the application of a small time window for the final lag search during flux calculations. 
[…]
